code audit fixes: mask type (float→bool), prosody normalization in pooling, mask-safe mean pooling
Fixes:
- collate: mask now bool tensor (not float32) for correct ~invert
- pooling: mask check uses `mask==0` instead of `~mask` for compatibility
- pooling: add per-batch F0/energy normalization (had raw Hz/RMS values)
- train: mask-aware mean pooling uses `mask.to(dtype)` for bool safety

Previously prosody features (F0 0-2000Hz, energy 0-0.5RMS) were fed raw
into the attention layer, causing F0 to dominate. Normalization should
yield more balanced F0+energy contributions.

+++ src/training/train_ssl.py
@@ -103,7 +103,7 @@ class DistributionCalibratedSER(nn.Module):
         else:
             # Mask-aware mean pooling
             if mask is not None:
-                mask_f = mask.float().unsqueeze(-1)  # (B, T, 1)
+                mask_f = mask.to(dtype=ssl_feats.dtype).unsqueeze(-1)  # (B, T, 1)
                 pooled = (ssl_feats * mask_f).sum(dim=1) / mask_f.sum(dim=1).clamp(min=1)
             else:
                 pooled = ssl_feats.mean(dim=1)
